feat: bottle hero
