+++ heroes/public/assets/SkullBottle.glb
